test: use default replication connection timeout
Removed all local connection timeout setups, where it is not the part
of the testing scenario. Instead of it the default value will be used.
Also it is more comfortable to change the single default value instead
of the change this value at each of the test scenarios, it will help
to avoid of flaky results in tests beacuse of different timeout vlaues.
The most of tests used the replication connection timeout is equal to
0.5 second, while defualt value is set at sources file:
    src/box/lua/load_cfg.lua
to the value of 30 seconds:
    replication_connect_timeout = 30
which is quite enough for all of the tests that don't need any special
values for it.

Part of #2058