popen: Introduce a backend engine
In the patch we introduce popen backend engine which provides
a way to execute external programs and communicate with their
stdin/stdout/stderr streams.

It is possible to run a child process with:

 a) completely closed stdX descriptors
 b) provide /dev/null descriptors to appropritate stdX
 c) pass new transport into a child (currently we use
    pipes for this sake, but mae extend to tty/sockets)
 d) inherit stdX from a parent, iow do nothing

On tarantool start we create @popen_pids_map hash which maps
created processes PIDs to popen_handle structure, this structure
keeps everything needed to control and communicate with the children.
The hash will allow us to find a hild process quickly from inside
of a signal handler.

Each handle links into @popen_head list, which is need to be able
to destory children processes on exit procedure (ie when we exit
tarantool and need to cleanup the resources used).

Every new process is born by vfork call - we can't use fork()
because of at_fork handlers in libeio which cause deadlocking
in internal mutex usage. Thus the caller waits until vfork finishes
its work and runs exec (or exit with error). The good news is that
we are supposed to create new processes from coio separate thread
thus other threads are not affected.

Because children processes are running without any limitations
they can exit by self or can be killed by some other third side
(say user of a hw node), we need to watch their state which is
done by setting a hook with ev_child_start helper. This helper
allows us to catch SIGCHLD when a child get exited/signaled
and unregister it from a pool or currently running children.
Note the libev wait() reaps child zomby by self. Such specifics
force us to use signal blocking procedure as a sync point to keep
children pids consistent (without signal blocking pids could be
simply reused inbetween and we can't distinguish the pid we're
tracking belongs our child process or some other process in
the system).

This engine provides the following API:
 - popen_create -- to create a new child process
 - popen_destroy -- to release resources occupied and
   terminate a chile process
 - popen_kill -- to kill a child process, note that this
   routine doesn't wait for process termination it simply
   sends SIGKILL signal
 - popen_wstatus -- to fetch current system dependant status
   of a child process
 - popen_stat -- to fetch statistic of a child process

Known issues:

 - environment variables are flushed to zero, should we provide
   a way to adjust it (via options) or inherit it instead?

 - popen_kill always send SIGKILL, should not we provide a
   portable way to customize signal sedning (say symbolic
   names for signals and pass them here)?

 - for native mode we don't do additional processing of arguments
   thus only plain name of elf executable will be working, we
   should provide a way for argv explicit passing or
   do analyze @command for arguments by hands;

 - for linux base systems we use popen2 system call passing
   O_CLOEXEC flag so that two concurrent popen_create calls
   would not affect each other with pipes inheritance (while
   currently we don't have a case where concurrent calls could
   be done as far as I know, still better to be on a safe side
   from the beginning);

 - there are some files (such as xlog) which tarantool opens
   for own needs without setting O_CLOEXEC flag and it get
   propagated to a children process; for linux based systems
   we use close_inherited_fds helper which walks over opened
   files of a process and close them but for other targets
   like MachO or FreeBSD this helper just zapped simply because
   I don't have such machines to experimant with; we should
   investigate this moment in more details later once base
   code is merged in;

 - need to consider a case where we will be using piping for
   descriptors (for example we might be writting into stdin
   of a child from another pipe, for this sake we could use
   splice() syscall which gonna be a way faster than copying
   data inside kernel between process). Still the question
   is -- do we really need it? Since we use interanal flags
   in popen handle this should not be a big problem to extend
   this interfaces.

Part of #4031

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>